[tapi] Mark some methods as weak to avoid duplicate symbols in Apple TAPI
Summary:
Apple TAPI 1100.0.11 (in external/libtapi) includes versions of these
six methods that cause duplicated symbols when linking both LLVMSupport
and TapiCore.

The three `ScalarTraits<FlowStringRef>` seems to be mostly identical in
both versions (the code is different, but visual inspection shows that
it will be equivalent, delegating on `ScalarTraits<StringRef>`).

Two of the `ScalarTraits<SwiftVersion>` are also identical (`output`
and `mustQuote`). However `ScalarTraits<SwiftVersion>::input` has extra
code that casts the provided context to some type that only exists in
LLVM, which will not work in Apple TAPI, so at least that method needs
to be avoided.

Instead of disabling 5 methods in the Apple TAPI 1100.0.11 codebase and
marking one method here as weak, mark all the methods here as weak to
have the hack localized only in one piece of code.

Test Plan: This works in my devserver in the ld64-609 branches I am creating. It should work the same without those branches too.

Reviewers: smeenai, leevince, #toolchain_core

Reviewed By: leevince

Subscribers: leevince

Differential Revision: https://phabricator.intern.facebook.com/D36138325

Tasks: T118130272

Tags: swift-adoption