Kuree's example that had the dangling lifetime
The problem is at

clang/lib/CIR/Lowering/DirectToLLVM/LowerToLLVM.cpp:838

The `mlir::ValueRange` constructor chosen is `std::initializer_list`
and thus the ilist is created on the stack at line 838 and passed to
the ctor. The ctor borrows the memory from the ilist and is stored in
a higher scope and thus stays around. 839 dtors the ilist and the
ValueRange is no longer pointing to living memory.